+++ src/devtools/devtools.ts
@@ -1,3 +1,7 @@
 import { getDefaultPlatformIdentity } from '../features/platform/platformIdentityRegistry';
 
-chrome.devtools.panels.create(getDefaultPlatformIdentity().panelName, '', 'panel.html');
+chrome.devtools.panels.create(
+  getDefaultPlatformIdentity().panelName,
+  'icons/icon-32.png',
+  'panel.html',
+);